backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
numpy==1.26.3
openai==0.28.0
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
tensorflow==2.15.0.post1